pyproject.toml
[tool.poetry]
name = "pdf-scraper"
version = "0.1.0"
description = ""
authors = ["Yossef-Dawoad <yossefdawoad15@gmail.com>"]
readme = "README.md"
packages = [{include = "pdf_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
tabula-py = "^2.8.1"
fastapi = {extras = ["all"], version = "^0.103.1"}
slowapi = "^0.1.8"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
ruff = "^0.0.290"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
